Data.Time.Conversion

Description

This module provides functions for reading time strings. We also provide functions for converting between timezones.

Since: 0.1

High-level parsing/conversion

readConvertTime :: Maybe TimeReader -> Maybe TZDatabase -> IO ZonedTime

Reads the given time string based on the source TimeReader and converts to the destination timezone. This is the composition of readTime and convertTime. If the source is Nothing then we read the local system time. Similarly, if the dest is Nothing, we convert to the local system timezone.

Throws:

ParseTimeException: Error parsing the time string.

ParseTZDatabaseException: Error parsing the tz database name.

LocalTimeZoneException: Error retrieving local timezone.

LocalSystemTimeException: Error retrieving local system time.

Examples

Expand

>>> let toUtc = Just $ TZDatabaseLabel Etc__UTC
>>> let reader = defaultTimeReader "17:24"
>>> let litReader = reader { srcTZ = Just SrcTZLiteral }
>>> -- literal + no src timezone = utc
>>> readConvertTime (Just litReader) toUtc
1970-01-01 17:24:00 UTC

>>> -- literal + convert from est
>>> readConvertTime (Just (litReader { format = Types.hmTZ, timeString = "17:24 EST" })) toUtc
1970-01-01 22:24:00 UTC

>>> let nyReader = litReader { srcTZ = Just (SrcTZDatabase $ TZDatabaseLabel America__New_York) }
>>> readConvertTime (Just (nyReader { timeString = "08:15" })) toUtc
1970-01-01 13:15:00 UTC

>>> let badTimeString = litReader { timeString = "bad" }
>>> readConvertTime (Just badTimeString) toUtc
*** Exception: MkParseTimeException {errFormat = MkTimeFormat {unTimeFormat = "%H:%M"}, errMsg = "bad"}

>>> let badTZDatabase = litReader { srcTZ = Just (SrcTZDatabase $ TZDatabaseText "America/NewYork")}
>>> readConvertTime (Just badTZDatabase) toUtc
*** Exception: MkParseTZDatabaseException {unParseTZDatabaseException = "America/NewYork"}

Since: 0.1

readTime :: Maybe TimeReader -> IO ZonedTime

Reads a time based on the TimeReader. If given Nothing we read the local system time instead. If srcTZ is SrcTZDatabase then we manually append the timezone data onto the time string and format. That is, only use SrcTZDatabase if:

timeString does not contain a timezone like EST.

format does not mention the timezone option %Z.

Throws:

ParseTimeException: Error parsing the time string.

ParseTZDatabaseException: Error parsing the tz database name.

LocalTimeZoneException: Error retrieving local timezone.

LocalSystemTimeException: Error retrieving local system time.

Examples

Expand

>>> let reader = defaultTimeReader "17:24"
>>> let litReader = reader { srcTZ = Just SrcTZLiteral }
>>> -- literal + no src timezone = utc
>>> readTime (Just litReader)
1970-01-01 17:24:00 +0000

>>> -- literal + convert from est
>>> readTime (Just (litReader { format = Types.hmTZ, timeString = "17:24 EST" }))
1970-01-01 17:24:00 EST

>>> let nyReader = litReader { srcTZ = Just (SrcTZDatabase $ TZDatabaseLabel America__New_York) }
>>> readTime (Just (nyReader { timeString = "08:15" }))
1970-01-01 08:15:00 EST

>>> let badTimeString = litReader { timeString = "bad" }
>>> readTime (Just badTimeString)
*** Exception: MkParseTimeException {errFormat = MkTimeFormat {unTimeFormat = "%H:%M"}, errMsg = "bad"}

>>> let badTZDatabase = litReader { srcTZ = Just (SrcTZDatabase $ TZDatabaseText "America/NewYork")}
>>> readTime (Just badTZDatabase)
*** Exception: MkParseTZDatabaseException {unParseTZDatabaseException = "America/NewYork"}

Since: 0.1

convertTime :: ZonedTime -> Maybe TZDatabase -> IO ZonedTime

Converts the given time to the destination timezone. If no destination timezone is given then we convert to the local system timezone.

Throws:

ParseTZDatabaseException: Error parsing the tz database name.

LocalTimeZoneException: Error retrieving local timezone.

Examples

Expand

>>> import Data.Time.Clock (UTCTime (..))
>>> import Data.Time.LocalTime (utc, utcToZonedTime)
>>> let zoned = utcToZonedTime utc (UTCTime (toEnum 50_000) 0)
>>> zoned
1995-10-10 00:00:00 UTC

>>> let toNy = Just $ TZDatabaseText "America/New_York"
>>> convertTime zoned toNy
1995-10-09 20:00:00 EDT

>>> convertTime zoned (Just $ TZDatabaseText "America/NewYork")
*** Exception: MkParseTZDatabaseException {unParseTZDatabaseException = "America/NewYork"}

Since: 0.1

Types

data TimeReader

Determines how to read a time string.

Since: 0.1

Constructors

MkTimeReader

Fields

format :: TimeFormat

Format used when parsing the time string.

Since: 0.1

srcTZ :: Maybe SrcTZ

Timezone in which to read the string. Nothing corresponds to local timezone. If srcTZ is SrcTZDatabase then:

timeString should not contain a timezone like EST.

format should not mention the timezone option %Z.

Since: 0.1

locale :: TimeLocale

Locale to use when parsing.

Since: 0.1

today :: Bool

True if this is the current day. This is a convenient way to parse the string as if it includes the current day's date.

Since: 0.1

timeString :: Text

The time string to parse.

Since: 0.1

defaultTimeReader :: Text -> TimeReader

Given a time string, returns a default time reader.

format = "%H:%M": 24hr hours:minutes

srzTZ = Nothing: local

locale = timeLocaleAllZones: all locales

today = False: do not automatically add current day

Since: 0.1

data SrcTZ

Determines what timezone we use when reading a time string.

Since: 0.1

Constructors

SrcTZDatabase TZDatabase
Use the TZDatabase to determine which timezone to use.

Since: 0.1

SrcTZLiteral
Timezones are not given any extra consideration. This is useful when either the string itself contains a timezone (e.g. "15:03 EST") or we want UTC.

Since: 0.1

data TZDatabase

Options for interpreting a timezone.

Examples

Expand

>>> TZDatabaseLabel America__New_York
TZDatabaseLabel America__New_York

>>> TZDatabaseText "America/New_York"
TZDatabaseText "America/New_York"

Since: 0.1

Constructors

TZDatabaseLabel TZLabel
Uses a TZLabel. This option should be preferred as it enables total conversions.

Since: 0.1

TZDatabaseText Text
Interprets the Text as a tz database name e.g. America/New_York. Obviously such conversions cannot be total.

Since: 0.1

newtype TimeFormat

Time formatting string. The Monoid instance behaves like Text, whereas Default is an alias for hm.

Examples

Expand

>>> def :: TimeFormat
MkTimeFormat {unTimeFormat = "%H:%M"}

>>> mempty :: TimeFormat
MkTimeFormat {unTimeFormat = ""}

>>> hm <> " %Z"
MkTimeFormat {unTimeFormat = "%H:%M %Z"}

Since: 0.1

Constructors

MkTimeFormat

Fields

unTimeFormat :: Text

Since: 0.1

Formatting

hm :: TimeFormat

Format for 24-hour hours:minutes.

>>> hm
MkTimeFormat {unTimeFormat = "%H:%M"}

Since: 0.1

hm12h :: TimeFormat

Format for 12-hour hours:minutes am/pm.

>>> hm12h
MkTimeFormat {unTimeFormat = "%I:%M %P"}

Since: 0.1

hmTZ :: TimeFormat

Format for 24-hour hours:minutes TZ.

>>> hmTZ
MkTimeFormat {unTimeFormat = "%H:%M %Z"}

Since: 0.1

hmTZ12h :: TimeFormat

Format for 12-hour hours:minutes am/pm TZ.

>>> hmTZ12h
MkTimeFormat {unTimeFormat = "%I:%M %P %Z"}

Since: 0.1

rfc822 :: TimeFormat

Format for RFC822: %a, %_d %b %Y %H:%M:%S %Z.

>>> rfc822
MkTimeFormat {unTimeFormat = "%a, %_d %b %Y %H:%M:%S %Z"}

Since: 0.1

Low-level functions

Parsing time strings

readInLocalTimeZone :: TimeLocale -> TimeFormat -> Text -> IO ZonedTime

readInLocalTimeZone locale format timeStr attempts to parse the timeStr given the expected format. We parse into the current system timezone, so:

format should not mention "%Z"

timeStr should not contain timezone information.

λ. readInLocalTimeZone Types.timeLocaleAllZones "%H" "17"
Just 1970-01-01 17:00:00 NZST

Throws:

ParseTimeException: Error parsing the time string.

LocalTimeZoneException: Error retrieving local timezone.

Since: 0.1

readTimeFormat :: TimeLocale -> TimeFormat -> Text -> Maybe ZonedTime

readTimeFormat locale format timeStr attempts to parse the timeStr given the expected format. No timezone is assumed, so if it is left off then the result is UTC.

Examples

Expand

>>> readTimeFormat Utils.timeLocaleAllZones Types.hm "17:24"
Just 1970-01-01 17:24:00 +0000

>>> readTimeFormat Utils.timeLocaleAllZones Types.hm12h "07:24 pm"
Just 1970-01-01 19:24:00 +0000

>>> readTimeFormat Utils.timeLocaleAllZones Types.hmTZ "07:24 CET"
Just 1970-01-01 07:24:00 CET

>>> readTimeFormat Utils.timeLocaleAllZones Types.hmTZ12h "07:24 pm EST"
Just 1970-01-01 19:24:00 EST

Since: 0.1

Converting ZonedTime

convertZoned :: ZonedTime -> Text -> Maybe ZonedTime

Converts a zoned time to the given timezone.

Examples

Expand

>>> let (Just sixPmUtc) = readTimeFormat Utils.timeLocaleAllZones Types.hm "18:00"
>>> convertZoned sixPmUtc "America/New_York"
Just 1970-01-01 13:00:00 EST

>>> convertZoned sixPmUtc "bad tz label"
Nothing

Since: 0.1

convertZonedLabel :: ZonedTime -> TZLabel -> ZonedTime

Converts a zoned time to the given timezone.

Examples

Expand

>>> let (Just sixPmUtc) = readTimeFormat Utils.timeLocaleAllZones Types.hm "18:00"
>>> convertZonedLabel sixPmUtc America__New_York
1970-01-01 13:00:00 EST

Since: 0.1

Errors

data ParseTimeException

Exception parsing time string.

Since: 0.1

Constructors

MkParseTimeException

Fields

errFormat :: TimeFormat

Since: 0.1

errMsg :: Text

Since: 0.1

newtype ParseTZDatabaseException

Exception parsing tz database names.

Since: 0.1

Constructors

MkParseTZDatabaseException

Fields

unParseTZDatabaseException :: Text

Since: 0.1

data LocalTimeZoneException

Exception reading local system timezone.

Since: 0.1

Constructors

forall e.Exception e => MkLocalTimeZoneException e

data LocalSystemTimeException

Exception reading local system time.

Since: 0.1

Constructors

forall e.Exception e => MkLocalSystemTimeException e

Optics

_SrcTZDatabase :: Prism' SrcTZ TZDatabase

Since: 0.1

_SrcTZLiteral :: Prism' SrcTZ ()

Since: 0.1

_TZDatabaseLabel :: Prism' TZDatabase TZLabel

Since: 0.1

_TZDatabaseText :: Prism' TZDatabase Text

Since: 0.1

_MkTimeFormat :: Iso' TimeFormat Text

Since: 0.1

_ParseTZDatabaseException :: Iso' ParseTZDatabaseException Text

Since: 0.1

_LocalTimeZoneException :: Exception e => Review LocalTimeZoneException e

Since: 0.1

_LocalSystemTimeException :: Exception e => Review LocalSystemTimeException e

Since: 0.1

Miscellaneous

timeLocaleAllZones :: TimeLocale

defaultTimeLocale with the date format switched to %d/%m/%y and knowledge of all timezones, per TZLabel. Using this, we can parse non-American labels like CES and NZST.

Since: 0.1

Reexports

data ZonedTime

A local time together with a time zone.

There is no Eq instance for ZonedTime. If you want to compare local times, use zonedTimeToLocalTime. If you want to compare absolute times, use zonedTimeToUTC.

Constructors

ZonedTime

Fields

zonedTimeToLocalTime :: LocalTime

zonedTimeZone :: TimeZone

data TZLabel

Enumeration of time zone locations.

Constructors

Africa__Abidjan

Africa__Algiers

Africa__Bissau

Africa__Cairo

Africa__Casablanca

Africa__Ceuta

Africa__El_Aaiun

Africa__Johannesburg

Africa__Juba

Africa__Khartoum

Africa__Lagos

Africa__Maputo

Africa__Monrovia

Africa__Nairobi

Africa__Ndjamena

Africa__Sao_Tome

Africa__Tripoli

Africa__Tunis

Africa__Windhoek

America__Adak

America__Anchorage

America__Araguaina

America__Argentina__Buenos_Aires

America__Argentina__Catamarca

America__Argentina__Cordoba

America__Argentina__Jujuy

America__Argentina__La_Rioja

America__Argentina__Mendoza

America__Argentina__Rio_Gallegos

America__Argentina__Salta

America__Argentina__San_Juan

America__Argentina__San_Luis

America__Argentina__Tucuman

America__Argentina__Ushuaia

America__Asuncion

America__Bahia

America__Bahia_Banderas

America__Barbados

America__Belem

America__Belize

America__Boa_Vista

America__Bogota

America__Boise

America__Cambridge_Bay

America__Campo_Grande

America__Cancun

America__Caracas

America__Cayenne

America__Chicago

America__Chihuahua

America__Costa_Rica

America__Cuiaba

America__Danmarkshavn

America__Dawson

America__Dawson_Creek

America__Denver

America__Detroit

America__Edmonton

America__Eirunepe

America__El_Salvador

America__Fort_Nelson

America__Fortaleza

America__Glace_Bay

America__Goose_Bay

America__Grand_Turk

America__Guatemala

America__Guayaquil

America__Guyana

America__Halifax

America__Havana

America__Hermosillo

America__Indiana__Indianapolis

America__Indiana__Knox

America__Indiana__Marengo

America__Indiana__Petersburg

America__Indiana__Tell_City

America__Indiana__Vevay

America__Indiana__Vincennes

America__Indiana__Winamac

America__Inuvik

America__Iqaluit

America__Jamaica

America__Juneau

America__Kentucky__Louisville

America__Kentucky__Monticello

America__La_Paz

America__Lima

America__Los_Angeles

America__Maceio

America__Managua

America__Manaus

America__Martinique

America__Matamoros

America__Mazatlan

America__Menominee

America__Merida

America__Metlakatla

America__Mexico_City

America__Miquelon

America__Moncton

America__Monterrey

America__Montevideo

America__New_York

America__Nipigon

America__Nome

America__Noronha

America__North_Dakota__Beulah

America__North_Dakota__Center

America__North_Dakota__New_Salem

America__Nuuk

America__Ojinaga

America__Panama

America__Pangnirtung

America__Paramaribo

America__Phoenix

America__Port_au_Prince

America__Porto_Velho

America__Puerto_Rico

America__Punta_Arenas

America__Rainy_River

America__Rankin_Inlet

America__Recife

America__Regina

America__Resolute

America__Rio_Branco

America__Santarem

America__Santiago

America__Santo_Domingo

America__Sao_Paulo

America__Scoresbysund

America__Sitka

America__St_Johns

America__Swift_Current

America__Tegucigalpa

America__Thule

America__Thunder_Bay

America__Tijuana

America__Toronto

America__Vancouver

America__Whitehorse

America__Winnipeg

America__Yakutat

America__Yellowknife

Antarctica__Casey

Antarctica__Davis

Antarctica__Macquarie

Antarctica__Mawson

Antarctica__Palmer

Antarctica__Rothera

Antarctica__Troll

Antarctica__Vostok

Asia__Almaty

Asia__Amman

Asia__Anadyr

Asia__Aqtau

Asia__Aqtobe

Asia__Ashgabat

Asia__Atyrau

Asia__Baghdad

Asia__Baku

Asia__Bangkok

Asia__Barnaul

Asia__Beirut

Asia__Bishkek

Asia__Brunei

Asia__Chita

Asia__Choibalsan

Asia__Colombo

Asia__Damascus

Asia__Dhaka

Asia__Dili

Asia__Dubai

Asia__Dushanbe

Asia__Famagusta

Asia__Gaza

Asia__Hebron

Asia__Ho_Chi_Minh

Asia__Hong_Kong

Asia__Hovd

Asia__Irkutsk

Asia__Jakarta

Asia__Jayapura

Asia__Jerusalem

Asia__Kabul

Asia__Kamchatka

Asia__Karachi

Asia__Kathmandu

Asia__Khandyga

Asia__Kolkata

Asia__Krasnoyarsk

Asia__Kuala_Lumpur

Asia__Kuching

Asia__Macau

Asia__Magadan

Asia__Makassar

Asia__Manila

Asia__Nicosia

Asia__Novokuznetsk

Asia__Novosibirsk

Asia__Omsk

Asia__Oral

Asia__Pontianak

Asia__Pyongyang

Asia__Qatar

Asia__Qostanay

Asia__Qyzylorda

Asia__Riyadh

Asia__Sakhalin

Asia__Samarkand

Asia__Seoul

Asia__Shanghai

Asia__Singapore

Asia__Srednekolymsk

Asia__Taipei

Asia__Tashkent

Asia__Tbilisi

Asia__Tehran

Asia__Thimphu

Asia__Tokyo

Asia__Tomsk

Asia__Ulaanbaatar

Asia__Urumqi

Asia__Ust_Nera

Asia__Vladivostok

Asia__Yakutsk

Asia__Yangon

Asia__Yekaterinburg

Asia__Yerevan

Atlantic__Azores

Atlantic__Bermuda

Atlantic__Canary

Atlantic__Cape_Verde

Atlantic__Faroe

Atlantic__Madeira

Atlantic__Reykjavik

Atlantic__South_Georgia

Atlantic__Stanley

Australia__Adelaide

Australia__Brisbane

Australia__Broken_Hill

Australia__Darwin

Australia__Eucla

Australia__Hobart

Australia__Lindeman

Australia__Lord_Howe

Australia__Melbourne

Australia__Perth

Australia__Sydney

Etc__GMT

Etc__GMT'1

Etc__GMT'10

Etc__GMT'11

Etc__GMT'12

Etc__GMT'2

Etc__GMT'3

Etc__GMT'4

Etc__GMT'5

Etc__GMT'6

Etc__GMT'7

Etc__GMT'8

Etc__GMT'9

Etc__GMT_1

Etc__GMT_10

Etc__GMT_11

Etc__GMT_12

Etc__GMT_13

Etc__GMT_14

Etc__GMT_2

Etc__GMT_3

Etc__GMT_4

Etc__GMT_5

Etc__GMT_6

Etc__GMT_7

Etc__GMT_8

Etc__GMT_9

Etc__UTC

Europe__Amsterdam

Europe__Andorra

Europe__Astrakhan

Europe__Athens

Europe__Belgrade

Europe__Berlin

Europe__Brussels

Europe__Bucharest

Europe__Budapest

Europe__Chisinau

Europe__Copenhagen

Europe__Dublin

Europe__Gibraltar

Europe__Helsinki

Europe__Istanbul

Europe__Kaliningrad

Europe__Kiev

Europe__Kirov

Europe__Lisbon

Europe__London

Europe__Luxembourg

Europe__Madrid

Europe__Malta

Europe__Minsk

Europe__Monaco

Europe__Moscow

Europe__Oslo

Europe__Paris

Europe__Prague

Europe__Riga

Europe__Rome

Europe__Samara

Europe__Saratov

Europe__Simferopol

Europe__Sofia

Europe__Stockholm

Europe__Tallinn

Europe__Tirane

Europe__Ulyanovsk

Europe__Uzhgorod

Europe__Vienna

Europe__Vilnius

Europe__Volgograd

Europe__Warsaw

Europe__Zaporozhye

Europe__Zurich

Indian__Chagos

Indian__Christmas

Indian__Cocos

Indian__Kerguelen

Indian__Mahe

Indian__Maldives

Indian__Mauritius

Indian__Reunion

Pacific__Apia

Pacific__Auckland

Pacific__Bougainville

Pacific__Chatham

Pacific__Chuuk

Pacific__Easter

Pacific__Efate

Pacific__Fakaofo

Pacific__Fiji

Pacific__Funafuti

Pacific__Galapagos

Pacific__Gambier

Pacific__Guadalcanal

Pacific__Guam

Pacific__Honolulu

Pacific__Kanton

Pacific__Kiritimati

Pacific__Kosrae

Pacific__Kwajalein

Pacific__Majuro

Pacific__Marquesas

Pacific__Nauru

Pacific__Niue

Pacific__Norfolk

Pacific__Noumea

Pacific__Pago_Pago

Pacific__Palau

Pacific__Pitcairn

Pacific__Pohnpei

Pacific__Port_Moresby

Pacific__Rarotonga

Pacific__Tahiti

Pacific__Tarawa

Pacific__Tongatapu

Pacific__Wake

Pacific__Wallis

Root__CET

Root__CST6CDT

Root__EET

Root__EST

Root__EST5EDT

Root__HST

Root__MET

Root__MST

Root__MST7MDT

Root__PST8PDT

Root__WET

data TimeLocale

Constructors

TimeLocale

Fields

wDays :: [(String, String)]

full and abbreviated week days, starting with Sunday

months :: [(String, String)]

full and abbreviated months

amPm :: (String, String)

AM/PM symbols

dateTimeFmt :: String

formatting strings

dateFmt :: String

formatting strings

timeFmt :: String

formatting strings

time12Fmt :: String

formatting strings

knownTimeZones :: [TimeZone]

time zones known by name