Data.Time.Conversion.Types

Description

This module provides types for reading / converting between timezones.

Since: 0.1

Documentation

data TimeReader

Determines how to read a time string.

Since: 0.1

Constructors

MkTimeReader

Fields

format :: TimeFormat

Format used when parsing the time string.

Since: 0.1

srcTZ :: Maybe SrcTZ

Timezone in which to read the string. Nothing corresponds to local timezone. If srcTZ is SrcTZDatabase then:

timeString should not contain a timezone like EST.

format should not mention the timezone option %Z.

Since: 0.1

locale :: TimeLocale

Locale to use when parsing.

Since: 0.1

today :: Bool

True if this is the current day. This is a convenient way to parse the string as if it includes the current day's date.

Since: 0.1

timeString :: Text

The time string to parse.

Since: 0.1

defaultTimeReader :: Text -> TimeReader

Given a time string, returns a default time reader.

format = "%H:%M": 24hr hours:minutes

srzTZ = Nothing: local

locale = timeLocaleAllZones: all locales

today = False: do not automatically add current day

Since: 0.1

data SrcTZ

Determines what timezone we use when reading a time string.

Since: 0.1

Constructors

SrcTZDatabase TZDatabase
Use the TZDatabase to determine which timezone to use.

Since: 0.1

SrcTZLiteral
Timezones are not given any extra consideration. This is useful when either the string itself contains a timezone (e.g. "15:03 EST") or we want UTC.

Since: 0.1

data TZDatabase

Options for interpreting a timezone.

Examples

Expand

>>> TZDatabaseLabel America__New_York
TZDatabaseLabel America__New_York

>>> TZDatabaseText "America/New_York"
TZDatabaseText "America/New_York"

Since: 0.1

Constructors

TZDatabaseLabel TZLabel
Uses a TZLabel. This option should be preferred as it enables total conversions.

Since: 0.1

TZDatabaseText Text
Interprets the Text as a tz database name e.g. America/New_York. Obviously such conversions cannot be total.

Since: 0.1

Formatting

newtype TimeFormat

Time formatting string. The Monoid instance behaves like Text, whereas Default is an alias for hm.

Examples

Expand

>>> def :: TimeFormat
MkTimeFormat {unTimeFormat = "%H:%M"}

>>> mempty :: TimeFormat
MkTimeFormat {unTimeFormat = ""}

>>> hm <> " %Z"
MkTimeFormat {unTimeFormat = "%H:%M %Z"}

Since: 0.1

Constructors

MkTimeFormat

Fields

unTimeFormat :: Text

Since: 0.1

hm :: TimeFormat

Format for 24-hour hours:minutes.

>>> hm
MkTimeFormat {unTimeFormat = "%H:%M"}

Since: 0.1

hm12h :: TimeFormat

Format for 12-hour hours:minutes am/pm.

>>> hm12h
MkTimeFormat {unTimeFormat = "%I:%M %P"}

Since: 0.1

hmTZ :: TimeFormat

Format for 24-hour hours:minutes TZ.

>>> hmTZ
MkTimeFormat {unTimeFormat = "%H:%M %Z"}

Since: 0.1

hmTZ12h :: TimeFormat

Format for 12-hour hours:minutes am/pm TZ.

>>> hmTZ12h
MkTimeFormat {unTimeFormat = "%I:%M %P %Z"}

Since: 0.1

rfc822 :: TimeFormat

Format for RFC822: %a, %_d %b %Y %H:%M:%S %Z.

>>> rfc822
MkTimeFormat {unTimeFormat = "%a, %_d %b %Y %H:%M:%S %Z"}

Since: 0.1

Errors

data ParseTimeException

Exception parsing time string.

Since: 0.1

Constructors

MkParseTimeException

Fields

errFormat :: TimeFormat

Since: 0.1

errMsg :: Text

Since: 0.1

newtype ParseTZDatabaseException

Exception parsing tz database names.

Since: 0.1

Constructors

MkParseTZDatabaseException

Fields

unParseTZDatabaseException :: Text

Since: 0.1

data LocalTimeZoneException

Exception reading local system timezone.

Since: 0.1

Constructors

forall e.Exception e => MkLocalTimeZoneException e

data LocalSystemTimeException

Exception reading local system time.

Since: 0.1

Constructors

forall e.Exception e => MkLocalSystemTimeException e

Optics

_SrcTZDatabase :: Prism' SrcTZ TZDatabase

Since: 0.1

_SrcTZLiteral :: Prism' SrcTZ ()

Since: 0.1

_TZDatabaseLabel :: Prism' TZDatabase TZLabel

Since: 0.1

_TZDatabaseText :: Prism' TZDatabase Text

Since: 0.1

_MkTimeFormat :: Iso' TimeFormat Text

Since: 0.1

_ParseTZDatabaseException :: Iso' ParseTZDatabaseException Text

Since: 0.1

_LocalTimeZoneException :: Exception e => Review LocalTimeZoneException e

Since: 0.1

_LocalSystemTimeException :: Exception e => Review LocalSystemTimeException e

Since: 0.1