Data.Time.Conversion.Types.TZDatabase

Description

Since: 0.1

Documentation

data TZDatabase

Options for interpreting a timezone.

Examples

Expand

>>> TZDatabaseLabel America__New_York
TZDatabaseLabel America__New_York

>>> TZDatabaseText "America/New_York"
TZDatabaseText "America/New_York"

Since: 0.1

Constructors

TZDatabaseLabel !TZLabel
Uses a TZLabel. This option should be preferred as it enables total conversions.

Since: 0.1

TZDatabaseText !Text
Interprets the Text as a tz database name e.g. America/New_York. Obviously such conversions cannot be total.

Since: 0.1

_TZDatabaseLabel :: Prism' TZDatabase TZLabel

Since: 0.1

_TZDatabaseText :: Prism' TZDatabase Text

Since: 0.1