Data.Time.Conversion.Types.TimeFormat

Description

Since: 0.1

Documentation

newtype TimeFormat

Time formatting string. The Monoid instance behaves like Text.

Examples

Expand

>>> defaultTimeFormat
MkTimeFormat {unTimeFormat = "%H:%M"}

>>> mempty :: TimeFormat
MkTimeFormat {unTimeFormat = ""}

>>> "%Y " <> hm
MkTimeFormat {unTimeFormat = "%Y %H:%M"}

Since: 0.1

Constructors

MkTimeFormat

Fields

unTimeFormat :: Text

defaultTimeFormat :: TimeFormat

Alias for hm.

hm :: TimeFormat

Format for 24-hour hours:minutes.

>>> hm
MkTimeFormat {unTimeFormat = "%H:%M"}

Since: 0.1

hm12h :: TimeFormat

Format for 12-hour hours:minutes am/pm.

>>> hm12h
MkTimeFormat {unTimeFormat = "%I:%M %P"}

Since: 0.1

hmTZ :: TimeFormat

Format for 24-hour hours:minutes TZ. As this contains a timezone flag, it should be used for formatting output only. In particular, it should not be used with TimeReader's format.

>>> hmTZ
MkTimeFormat {unTimeFormat = "%H:%M %Z"}

Since: 0.1

hmTZ12h :: TimeFormat

Format for 12-hour hours:minutes am/pm TZ. As this contains a timezone flag, it should be used for formatting output only. In particular, it should not be used with TimeReader's format.

>>> hmTZ12h
MkTimeFormat {unTimeFormat = "%I:%M %P %Z"}

Since: 0.1

rfc822 :: TimeFormat

Format for RFC822: %a, %_d %b %Y %H:%M:%S %Z. As this contains a timezone flag, it should be used for formatting output only. In particular, it should not be used with TimeReader's format.

>>> rfc822
MkTimeFormat {unTimeFormat = "%a, %_d %b %Y %H:%M:%S %Z"}

Since: 0.1